+++ v2/.gitignore
@@ -1,5 +1,7 @@
 ### Custom ###
-nextjs/**
+next
+
+.vscode
 
 environments/env.*.js
 !environments/env.example.js
@@ -11,6 +13,7 @@ environments/env.*.js
 .idea
 **/node_modules/*
 **/.firebaserc
+
 
 ### Firebase Patch ###
 .runtimeconfig.json
