+++ v2/firebase.json
@@ -6,7 +6,7 @@
     "source": ".",
     "ignore": [
       "firebase.json",
-      "firbease-debug.log",
+      "firebase-debug.log",
       "**/.*",
       "**/node_modules/**",
       "components/**",
