+++ v2/next.config.js
@@ -1,10 +1,11 @@
+const path = require('path');
 const withOptimizedImages = require('next-optimized-images');
 const env = require('./environments')(process.env.NODE_ENV || 'development');
 
 module.exports = withOptimizedImages({
-  distDir: 'nextjs',
+  distDir: 'next',
   env: {
-    FIREBASE_PROJECT_ID: 'gentlestudent-dev',
+    FIREBASE_PROJECT_ID: env.public.FIREBASE.projectId,
     FIREBASE: env.public.FIREBASE
   },
   experimental: {
