+++ v2/package.json
@@ -4,9 +4,9 @@
   "private": true,
   "main": "server.js",
   "scripts": {
-    "clean": "rimraf nextjs/ out/",
-    "dev": "cross-env NODE_ENV=development next dev",
-    "build": "cross-env NODE_ENV=production next build",
+    "clean": "rimraf build/ out/",
+    "dev": "cross-env NODE_ENV=development next dev src",
+    "build": "cross-env NODE_ENV=production next build src",
     "start": "next start",
     "local": "cross-env NODE_ENV=development firebase emulators:start",
     "predeploy": "npm run clean && next build && node ./utils/export.js",
@@ -19,13 +19,12 @@
     "@fortawesome/react-fontawesome": "^0.1.9",
     "dotenv": "^8.2.0",
     "firebase": "^7.14.2",
-    "firebase-admin": "^8.11.0",
-    "firebase-functions": "^3.6.1",
+    "firebase-admin": "^8.12.1",
+    "firebase-functions": "^3.6.2",
     "imagemin-mozjpeg": "^8.0.0",
     "imagemin-optipng": "^7.1.0",
     "isomorphic-unfetch": "^3.0.0",
     "next": "^9.3.5",
-    "next-optimized-images": "^2.6.0",
     "normalize.css": "^8.0.1",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
@@ -34,12 +33,11 @@
     "styled-jsx": "^3.2.5",
     "swr": "^0.2.0"
   },
-  "engines": {
-    "node": "10"
-  },
   "devDependencies": {
+    "@babel/core": "^7.10.2",
     "@babel/preset-env": "^7.9.6",
     "@babel/preset-react": "^7.9.4",
+    "@firebase/app-types": "^0.6.1",
     "babel-eslint": "^10.1.0",
     "cross-env": "^7.0.2",
     "eslint": "^6.8.0",
@@ -52,8 +50,10 @@
     "eslint-plugin-react-hooks": "^2.5.0",
     "firebase-functions-test": "^0.2.1",
     "next-images": "^1.4.0",
+    "next-optimized-images": "^2.6.1",
     "prettier": "^2.0.2",
     "rimraf": "^3.0.2",
-    "shelljs": "^0.8.4"
+    "shelljs": "^0.8.4",
+    "webpack": "^4.43.0"
   }
 }
