+++ v2/src/assets/styles/global.js
